验证/test1/backend/requirements.txt
# Python依赖列表
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.12.3
python-dotenv==1.0.0
pytest